refactor: 디렉토리 이름 수정
- abstract에서 base로 수정
