+++ src/js/components/Navbar/NavbarSub.js
@@ -1,4 +1,4 @@
-import { Component } from '../abstract/Component.js';
+import { Component } from '../base/Component.js';
 
 export class NavbarSub extends Component {
   constructor() {
