+++ src/js/components/base/Component.js
